Inkscape saved SVG files are parsed incorrect
Line count for location is wrong if character after name tag is `\n`. If it is a space/tab it is not part of the output any more (attribute after `data-inspect-column` is not separated anymore). Use look ahead assertion that does not consume this character. Add `use` as possible tag to locate.